refactor(api): products folder moved inside features folder
